Timer Scheduling Worked

+++ src/main/java/bupt/ugrd/model/Buginfo2Repository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
  * Created by Luyao on 2018/1/21.
  */
 
-public interface Buginfo2Repository extends CrudRepository<Buginfo2, Long> {
+public interface Buginfo2Repository extends CrudRepository<Buginfo2, Integer> {
 }
