+++ src/main/java/bupt/ugrd/model/BuginfoRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
  * Created by Luyao on 2018/1/21.
  */
 
-public interface BuginfoRepository extends CrudRepository<Buginfo, Long> {
+public interface BuginfoRepository extends CrudRepository<Buginfo, Integer> {
 }
