+++ src/main/java/bupt/ugrd/model/ContentRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Created by Luyao on 2018/1/21.
  */
-public interface ContentRepository extends CrudRepository<Content, Long> {
+public interface ContentRepository extends CrudRepository<Content, Integer> {
+
 }
