+++ src/main/java/bupt/ugrd/model/UserinfoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Created by Luyao on 2018/1/21.
  */
-public interface UserinfoRepository extends CrudRepository<Userinfo, Long> {
+public interface UserinfoRepository extends CrudRepository<Userinfo, Integer> {
 }
